requirements.txt
praw
supabase
google-generativeai
python-dotenv 